api/requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
starlette
pandas
numpy
psycopg2
tortoise-orm
aerich
asyncpg
passlib
bcrypt
python-jose
python-multipart